feat: 트랜잭션 고려해서 조회수 증가 로직 분리

+++ src/main/java/com/umc/product/notice/adapter/in/web/dto/response/query/GetNoticeDetailResponse.java
@@ -29,7 +29,7 @@ public record GetNoticeDetailResponse(
     NoticeTargetInfo targetInfo,
 
     // 메타데이터
-    Integer viewCount,
+    Long viewCount,
     Instant createdAt
 ) {
     public static GetNoticeDetailResponse from(
